Changed the PlayerEffect/_PlayerEffect interaction
PlayerEffect now contains the on_f and off_f functions, and passes the
instance of itself to the _PlayerEffect instance, instead of creating
silly classes inside of its `__init__`. This also allowed proper
`on/off` decorators, returning a new instance of `PlayerEffect` instead
of modifying the old one, allowing subclassing.